Break out material animation class and break the parent link.
